requirements.txt
requests==2.28.1
PyYAML
configparser==5.0.0
python-dateutil==2.8.0
simplejson==3.17.2
ruamel.yaml==0.16.12
urllib3~=1.26.16
curlify~=2.2.1

pytest==7.2.0
pytest-sugar
pytest-xdist==2.1.0
pytest-rerunfailures==11.1.1
pytest-custom-exit-code
pytest-html
pytest-flakes
pytest-pep8
flake8

allure_python_commons==2.11.1
allure-pytest
